kosmos-reference/kosmos-karpathy/pyproject.toml
[project]
name = "karpathy"
version = "0.1.0"
description = "Karpathy: An agentic Machine Learning Engineer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.6",
    "google-adk>=1.18.0",
    "litellm>=1.79.3",
    "markitdown[all]>=0.1.3",
    "mcp>=1.21.1",
    "modal>=1.2.2",
    "openai>=2.8.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]
